requirements.txt
opencv-python
pylint
pytest
pytest-cov
black
Pillow  
colour_demosaicing
matplotlib
seaborn
progressbar   
pandas
scikit-video  
PyYAML
openpyxl
pylint-exit